+++ packages/integration-sdk-core/src/data/__tests__/createIntegrationEntity.test.ts
@@ -61,6 +61,40 @@ describe('createIntegrationEntity', () => {
     expect(entity).toHaveProperty('active', true);
   });
 
+  test.each(['id', 'providerId'])(
+    'requires _key when %s is not a single string value',
+    (prop) => {
+      expect(() =>
+        createIntegrationEntity({
+          entityData: {
+            assign: networkAssigns,
+            source: {
+              ...networkSourceData,
+              [prop]: ['yodog', 1234],
+            },
+          },
+        }),
+      ).toThrow(/as type string/);
+
+      const entity = createIntegrationEntity({
+        entityData: {
+          assign: {
+            ...networkAssigns,
+            _key: 'assigned-because-id-isnt-usable',
+          },
+          source: {
+            ...networkSourceData,
+            [prop]: ['yodog', 1234],
+          },
+        },
+      });
+
+      expect(entity).toMatchObject({
+        _key: 'assigned-because-id-isnt-usable',
+      });
+    },
+  );
+
   test('should assign createdOn timestamp from creationDate', () => {
     const entity = createIntegrationEntity({
       entityData: {